feat: handle record not found error in RemoveAssignorUseCase

+++ backend/src/assignor/usecases/remove-assignor-usecase.ts
@@ -2,12 +2,19 @@ import { Injectable } from '@nestjs/common';
 import { RemoveAssignorInputDTO } from '../dto/remove-assignor.input.dto';
 import { IAssignorRepository } from '../repositories/assignor.repository.interface';
 import { IRemoveAssignorUseCase } from './remove-assignor.usecase.interface';
+import { RecordNotFoundError } from 'src/persistence/errors/record-not-found.error';
 
 @Injectable()
 export class RemoveAssignorUseCase implements IRemoveAssignorUseCase {
   constructor(private readonly assignorRepository: IAssignorRepository) {}
 
-  execute(removeAssignorDTO: RemoveAssignorInputDTO): Promise<void> {
+  async execute(removeAssignorDTO: RemoveAssignorInputDTO): Promise<void> {
+    const assignor = await this.assignorRepository.findById(
+      removeAssignorDTO.id,
+    );
+    if (!assignor) {
+      throw new RecordNotFoundError('Assignor');
+    }
     return this.assignorRepository.remove(removeAssignorDTO.id);
   }
 }
